dev: [Test] Update tests.
