+++ tests/DiContainer/Has/Fixtures/ClassWithSimpleDependency.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\DiContainer\Has\Fixtures;
+
+class ClassWithSimpleDependency
+{
+    public function __construct(private string $dependency) {}
+}
