+++ tests/DiContainer/Has/Fixtures/MyInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\DiContainer\Has\Fixtures;
+
+interface MyInterface {}
