--- tests/Fixtures/RuleInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Fixtures;
-
-interface RuleInterface {}
